Finished the backend work for add/remove an employee/exhibition

+++ hsr_museum/src/main/model/structures/Events.cs
@@ -9,11 +9,11 @@ namespace hsr_museum.src.main.model.structures
     public class Events
     {
         [JsonProperty("ID")]
-        public ulong id { get; private set; }
+        public uint id { get; private set; }
         [JsonProperty("Name")]
         public string name { get; private set; }
         
-        public Events(ulong ID, string Name) {
+        public Events(uint ID, string Name) {
             this.id = ID;
             this.name = Name;
         }
